修复参数互相影响问题：类型不匹配导致的内存越界
**问题根源：**
- image.cpp 中动态前瞻参数声明为 uint8（1字节）
- 但 auto_menu.cpp 中注册为 TYPE_UINT16（2字节）
- 修改参数时写入2字节到1字节空间，覆盖下一个参数内存

**修复内容：**

1. **修改参数类型定义（image.cpp + image.h）**
   - dynamic_weight_enable: uint8 → uint16
   - curvature_threshold_low: uint8 → uint16
   - curvature_threshold_high: uint8 → uint16
   - weight_shift_speed: uint8 → uint16
   - 保持 curvature_filter_ratio 为 float（菜单注册正确）

2. **去掉 CTRL 页面（auto_menu.cpp）**
   - 从 7 个页面减少到 6 个
   - 页面列表：SERVO, MOTOR, DIFFER, IMAGE, DYNAMIC, FUNC

3. **恢复 config_save() 调用**
   - 类型匹配后，自动保存功能安全恢复
   - 修改参数后立即保存到 Flash

**修复后效果：**
- DYNAMIC 页面参数调整不再互相影响
- 所有参数类型与菜单注册类型一致
- 内存安全，无越界写入

+++ C_Car_project/loongson_project/code/image.cpp
@@ -39,10 +39,10 @@ uint16 mid_weight_select = 2;   // 权重数组选择（1-5，默认2）
 uint16 cross_enable = 0;        // 十字识别开关（默认关闭）
 
 // ==================== 动态前瞻权重配置参数 ====================
-uint8 dynamic_weight_enable = 1;    // 动态权重开关（0=关闭使用固定权重, 1=开启）
-uint8 curvature_threshold_low = 5;  // 曲率低阈值（小于此值为直道，远前瞻）
-uint8 curvature_threshold_high = 15;// 曲率高阈值（大于此值为急弯，近前瞻）
-uint8 weight_shift_speed = 3;       // 权重切换速度（1-10，值越大切换越快）
+uint16 dynamic_weight_enable = 1;    // 动态权重开关（0=关闭使用固定权重, 1=开启）
+uint16 curvature_threshold_low = 5;  // 曲率低阈值（小于此值为直道，远前瞻）
+uint16 curvature_threshold_high = 15;// 曲率高阈值（大于此值为急弯，近前瞻）
+uint16 weight_shift_speed = 3;       // 权重切换速度（1-10，值越大切换越快）
 float curvature_filter_ratio = 0.7; // 曲率滤波系数（0-1，越大越平滑）
 
 uint8 reference_point = 0;
